+++ src/Tgstation.Server.Host/Controllers/RepositoryController.cs
@@ -153,7 +153,7 @@ namespace Tgstation.Server.Host.Controllers
 					Description = String.Format(CultureInfo.InvariantCulture, "Clone branch {1} of repository {0}", origin, cloneBranch ?? "master"),
 					StartedBy = AuthenticationContext.User,
 					CancelRightsType = RightsType.Repository,
-					CancelRight = (int)RepositoryRights.CancelClone,
+					CancelRight = (ulong)RepositoryRights.CancelClone,
 					Instance = Instance
 				};
 				var api = currentModel.ToApi();
@@ -307,7 +307,7 @@ namespace Tgstation.Server.Host.Controllers
 				StartedBy = AuthenticationContext.User,
 				Instance = Instance,
 				CancelRightsType = RightsType.Repository,
-				CancelRight = (int)RepositoryRights.CancelPendingChanges,
+				CancelRight = (ulong)RepositoryRights.CancelPendingChanges,
 			};
 
 			await jobManager.RegisterOperation(job, async (paramJob, serviceProvider, progressReporter, ct) =>
